requirements.txt
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.3
openai>=1.54.0
pyyaml>=6.0.2
python-slugify>=8.0.4
python-dateutil>=2.9.0
python-dotenv>=1.0.1
tenacity>=8.2.3